fix(anta.tests): revert code linting and reduce to VerifyISISSegmentRoutingTunnels only
